Fix logic that delete "-wal" file.
"databaseList()" method, All files in database folder are returned.
Such as "-wal" file, "-shm" file and so on.

In case of media provider, it uses WAL mode.
If "-wal" file is deleted, side effects appear to SQLITE_IOERR_SHORT_READ and user data is deleted.
Therefore, When deleting a database file, ".db" extension should be checked.

Signed-off-by : kwangjung kim <kwangjung.kim@lge.com>

Change-Id: Ie833d3990f1988929665d4c1b2ea77ad92bd2ded